refactor: move GrpcOptions to servers/grpc
