+++ src/frontend/src/service_config.rs
@@ -13,7 +13,6 @@
 // limitations under the License.
 
 pub mod datanode;
-pub mod grpc;
 pub mod influxdb;
 pub mod mysql;
 pub mod opentsdb;
@@ -21,7 +20,6 @@ pub mod otlp;
 pub mod postgres;
 pub mod prom_store;
 
-pub use grpc::GrpcOptions;
 pub use influxdb::InfluxdbOptions;
 pub use mysql::MysqlOptions;
 pub use opentsdb::OpentsdbOptions;
